Use float values for vertex coordinates
- Changed all functions using vertex coordinates to use floats instead
  of integers
- Scale z to be from 1 to 128 instead of 0 to screen width

+++ framebuffer.h
@@ -14,7 +14,7 @@ typedef struct color {
 } color;
 
 typedef struct vertex {
-	int x, y, z;
+	float x, y, z;
 } vertex;
 
 void init_framebuffer(fbuf *buf, int bufwidth, int bufheight);
